# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "prempawar130@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addGroup": {
            "post": {
                "tags": [
                    "Project Group"
                ],
                "summary": "Create Group",
                "description": "create group<br/>\n       Success Code:<br/>\n       10007: Group added successfully.<br/>\n       Error Code:<br/>\n       7002: Please enter a valid session-token.<br/>\n       7009: Please enter a valid group_name.<br/>\n       7010: Please enter a valid group_type(PUBLIC/PRIVATE).<br/>\n       7011: Group name is already exists.<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\groupController::addGroup",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide Group details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grp_name",
                                    "grp_note",
                                    "grp_type"
                                ],
                                "properties": {
                                    "grp_name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "grp_note": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "grp_type": {
                                        "type": "string",
                                        "example": "PUBLIC/PRIVATE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/deleteGroup": {
            "post": {
                "tags": [
                    "Project Group"
                ],
                "summary": "Delete Group",
                "description": "create group<br/>\n       Success Code:<br/>\n       10008: Group deleted successfullly.<br/>\n       Error Code:<br/>\n       7002: Please enter a valid session-token.<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\groupController::deleteGroup",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide Group details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grp_id"
                                ],
                                "properties": {
                                    "grp_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updateGroup": {
            "post": {
                "tags": [
                    "Project Group"
                ],
                "summary": "update Group",
                "description": "update group<br/>\n       Success Code:<br/>\n       10009: Group details update successfull.<br/>\n       Error Code:<br/>\n       7002: Please enter a valid session-token.<br/>\n       7009: Please enter a valid group_name.<br/>\n       7010: Please enter a valid group_type(PUBLIC/PRIVATE).<br/>\n       7011: Group name is already exists.<br/>\n       7012: Please enter a valid gorup_id.<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\groupController::updateGroup",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide Group details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grp_id",
                                    "grp_name",
                                    "grp_note",
                                    "grp_type"
                                ],
                                "properties": {
                                    "grp_id": {
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "grp_name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "grp_note": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "grp_type": {
                                        "type": "string",
                                        "example": "PUBLIC/PRIVATE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/listAll": {
            "get": {
                "tags": [
                    "Project Group"
                ],
                "summary": "List all folder names under usr_id",
                "description": "fetch folder list <br/>\n      Success Code:<br/>\n            10010: Group list fetched successfullly.<br/>\n      Error Code:<br/>\n            7002: Please enter a valid session-token.<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\groupController::fetchAllGroup",
                "parameters": [
                    {
                        "name": "grp_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "PUBLIC/PRIVATE"
                        }
                    },
                    {
                        "name": "num",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "10/50/150"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "0"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "grp_name/grp_note"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ASC/DESC"
                        }
                    },
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/details": {
            "get": {
                "tags": [
                    "Project Group"
                ],
                "summary": "List all group names under usr_id",
                "description": "fetch group list <br/>\n      Success Code:<br/>\n            10011: Group derails fetch successfully.<br/>\n      Error Code:<br/>\n            7002: Please enter a valid session-token.<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\groupController::groupDetails",
                "parameters": [
                    {
                        "name": "grp_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "numeric",
                            "example": "0"
                        }
                    },
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/swa_parameter": {
            "get": {
                "tags": [
                    "PREM Tag"
                ],
                "summary": "Get list of available subscription plans for user",
                "description": "Subscription list",
                "operationId": "App\\Http\\Controllers\\parametercontroller::parameter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}